LayoutTests/http/tests/navigation/frames-same-url-cancel-load-expected.txt

Test that when several frames are loading the same URL, and you navigate one, it does not cancel the load for the other frames.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS frame2.contentDocument.URL is "http://127.0.0.1:8000/navigation/resources/success.html"
PASS successfullyParsed is true

TEST COMPLETE
